Fix deno in build script
As of the time of this commit, deno is not available in v3.18, and is only
available in the community edge channel. Until it's available in the main
channel and can be put in the packages list, the deno command will need to be
manually added with a apk command.